# src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.0'
  title: Care Platform
  description: 'This is the API definition for the Care Platform.

  Authentication  Nodes:
  Click Authorize to enter User and Component authorization codes.

  - userAuth: The JWT token provided by the login command. Since the swagger 2.0 (OpenAPI) specification does
       not natively support bearer authentication. Please use this APIKey field. Prefix it with the word Bearer,
       like this: Bearer JWT_TOKEN
  - componentAuth: The key of the component, this can be obtained by making GET requests to .../components.'
host: localhost:10010
basePath: /v0
schemes:
- http
#s
consumes:
- application/json
produces:
- application/json
securityDefinitions:
   userAuth:
     type: apiKey
     in: header
     name: X-API-Key

   componentAuth:
     type: apiKey
     in: header
     name: X-API-Key
paths:
...
  /clients/{clientID}/components:
    x-swagger-router-controller: components
    post:
      security:
      - userAuth: []
      description: Create a new component for this client, generating a key to be given to the component. 
      summary:  Register a new component
      tags:
      - client
      operationId: newComponent
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to which the component should be added
      - name: body
        in: body
        required: true
        description: New component information
        schema:
          $ref: '#/definitions/NewComponent'
      responses:
        201:
          description: The component was created
          schema:
            $ref: '#/definitions/Component'
        401:
          description: Invalid credentials supplied
        404:
          description: The client does not exist
          
    get:
      security:
      - userAuth: []
      description: See the components associated with this client 
      summary:  Get client's components
      tags:
      - client
      operationId: getComponents
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Component'
        401:
          description: Invalid credentials supplied
        404:
          description: The client does not exist
...
  /clients/{clientID}/components/{componentID}/permissions:
    x-swagger-router-controller: permissions
    get:
      security:
      - userAuth: []
      description: If this compoent has any accepted permssions, you can get them here.
      summary:  See the permissions the component has been granted.
      tags:
      - client
      operationId: getPermissions
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to get
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Permission'
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist, or you are not the client's admin.
  /clients/{clientID}/components/{componentID}/permissions/requested:
    x-swagger-router-controller: permissions
    get:
      security:
      - userAuth: []
      description: If the component has requested any new permissions, you can get them here.
      summary:  See the permissions the component has requested
      tags:
      - client
      operationId: getRequestedPermissions
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to get
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Permission'
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist, or you are not the client's admin.
  /clients/{clientID}/components/{componentID}/permissions/accept:
    x-swagger-router-controller: permissions
    post:
      security:
      - userAuth: []
      description: Accept permissions that where requested by the component
      summary:  Accept permissions that where requested by the component
      tags:
      - client
      operationId: acceptPermissions
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to get
      - name: body
        in: body
        required: true
        description: List of permissions to accept
        schema:
            type: array
            items:
              $ref: '#/definitions/Permission'
      responses:
        200:
          description: Permissions were accepted
        400:
          description: Permission(s) not requested or already accepted
          schema:
              type: array
              items:
                $ref: '#/definitions/Permission'
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist, or you are not the client's admin.
  /clients/{clientID}/components/{componentID}/permissions/revoke:
    x-swagger-router-controller: permissions
    post:
      security:
      - userAuth: []
      description: Revoke previously accepted permissions.
      summary:  Revoke previously accepted permissions
      tags:
      - client
      operationId: revokePermissions
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to get
      - name: body
        in: body
        required: true
        description: List of permissions to revoke
        schema:
            type: array
            items:
              $ref: '#/definitions/Permission'
      responses:
        200:
          description: Permissions were revoked
        400:
          description: Permission were not accepted anyway
          schema:
              type: array
              items:
                $ref: '#/definitions/Permission'
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist, or you are not the client's admin.
...
  /components/requestPermissions:
    x-swagger-router-controller: permissions
    post:
      security:
      - componentAuth: []
      description: Request permissions. You should specify device defintions for any device permissions required. 
      summary:  Request permissions and define new devices
      tags:
      - component
      operationId: requestPermissions
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: List of permissions and defitions for any devices you wish to use
        schema:
          type: object
          properties:
            permissions:
              type: array
              items:
                $ref: '#/definitions/Permission'
            deviceDefinitions:
              type: array
              items:
                $ref: '#/definitions/DeviceDefinition'
      responses:
        200:
          description: Permissions were sent
        400:
          description: Device already exists with a different prototype
        401:
          description: The component key you supplied is invalid
...
definitions:
...
  NewComponent:
    title: Component to create
    type: object
    properties:
      type:
        type: string
        enum: 
        - event
        - model
        example: event
      name: 
        type: string
        example: Pill Box
      apiEndPoint:
        type: string
        example: pillboxexample.com/api
    required:
    - name
    - type
  Component:
    title: Component
    type: object
    allOf:
    - $ref: '#/definitions/NewComponent'
    properties:
      id:
        type: integer
        example: 0
      key:
        type: string
        example: pJN2GqumRbOmtNuCgKlhxjDgRWepazFGuCiikGcizc3GtUM9d8MXz5LCfEI0DjS0HzeiWODd9VGCjcTenttVnsCxG4RH1Sj1QMHe
    required:
    - id
    - key
    - type
  Permission:
    type: object
    title: Component Permission
    properties:
      type:
        type: string
        enum:
        - event
        - device
        - alert
        example: 'device'
      name:
        type: string
        example: "Pill Box"
    required:
    - type
    - name
  DeviceDefinition:
    type: object
    title: Device Definition
    properties:
      type:
        type: string
        example: Pill Box
      prototype:
        type: string
        example: "{\"location\": \"string\"}"
    required:
    - type
    - prototype
...